Bumped version number to 0.4 for development of this version.
Have added a new user flag called 'muted'. After thinking that
banning users is not a good idea on a system where the data belongs
to the user, muting is now suggested.

Muting ONLY STOPS posting and commenting. This should stop users
causing harm to any other users. It does no stop them from voting,
following, or gaining new followers.

We need to have a method in place to limit this for a set period of time.

Banning has not been removed until we know for a fact there is no time
this should be used.