Added data dumps. The can be accessed from the settings menu :)
This ONLY dumps account, posts and comments which YOU have made. This is
not an API.

User IDs (uids) have been replaced with <UID> and password hashes have been
replaced with <PASSWORD HASH>. These are the only thing I can remotely
think would be a security risk, but they probably aren't.

All other data such as votes on posts are not the users property. I would
also not want to dump the zset of uids.

It will be expanded more as and when we add more stuff to the site :D

Comment deleting where a user does not know where the original post is
also needs to be thought about. This could highlight some inaccessible ones.