Added post subscription feature.
This is to get ready for alerts. At the moment there is 3 ways to be subscribed to a post.

1 = You are the post creator
2 = You commented on the post
3 = You were taged in the post or a comment on the post

There is also a template tag to check subscription status and a unsubscribe view which at the moment is triggered by an action-button on the view post view.

There are some things which will be moved to Hungry Otter once development on it starts.