Fixes an issue with form error handling.
The way the error handling for dodgy posts is handled displays the first error. In a strange issue which I can not recreate, the check for errors returns true but there is no error. This means errors[0] fails. Added a quick check for the length of the list. In the case of one of these errors nothing will be displayed.