Removed some alerts stuff.
Alerts requires some more thinking. Little stubs such as the alerts view will probably be the same but lib/alerts.py requires a better design. This commit added some things but also reset a few.

Subscribtion reasons have been moved in to posts as this is a complete posts feature.

Also added ALERT_ITEMS_PER_PAGE to settings so that more can be shown compared to posts and comments.