Add subscription_reason function to posts.backend.
The new function simply takes a uid and a pid and returns the reason. Returning None if there is no subscription reason.

Note: Also removed alert keys from lib.keys. We are still working on a good design for alerts. The users will always have an alert so user:$uid:alerts has stayed.