Removed alerts code for now!
Alerts needs re-thinking. In working on it I noticed a few things that could be carried on in the rest of the system to make a notifications system seem less alien in the code base. This is being worked on but I would like the ability to change things so I'm going to remove what is here for alerts. The only thing left is the template which does nothing and a very basic view in users.