Alerts.
They are still not finished but there has been a lot of work done.

Alerts have a base class BaseAlert and an AlertManager which is responsible for loading, dumping and distributing them. These live in lib.alerts

Each module which uses alert subclass this and create there own alerts. At the moment alerts are only generated for following, tagging and commenting on a post you are subscribed too (see earlier commits in 0.4)

Alerts are stored as base64 encoded json-serialisation. This means we can possibly unpick them if we ever need to move from Python. This relies on the jsonpickle package.

Alerts are accessed from the top menu, these can only be accessed by the authecticated user @ /alerts

There are also prelim backend unit tests which are passing.

As I said this is not finished. I have added in a stub view call /i-has-alerts (name will change, probably) so we can get a boolean value of if we have alerts or not. This may not be implemented in 0.4 but will allow us to dynamically check for alerts whithout a page refresh.

There is also work that needs doing to highlight when there are new alerts, the templates need cleaning up and better unit testing. All still too come :)