ALerts, almost there.
Done a bit more work. Now have an endpoint which returns a JSON result if there are any new ones (may need changing to return a response jQuery can parse easily)

Added the same thing as a templatetag so the users with JS can still play.

Also double checked a few things with the Redis documents.

More tests still need writing, but we are on the right tracks.