Removed wsgi.py from coverage.
This can not really be tested an there would need to be a seperate test case and probably suite due to application contexts and what not.

I have already removed */tests.py as this may skew the results because obviously they will always be tested. So I hope wsgi.py being removed is not classed as cheating, it is a simply, simple file.