Refactor for 0.5 started!
For the 0.5 release of Pjuu we have decided to refactor the code base in nearly its entirety.

We are doing this as we can squash any potential issues, coding mistakes or design decisions we have made up to this point. It will the site is more secure and allow us to add new features.

There is a lot in this initial commit:

- ALL KEYS ARE NOW USE HASH TAGS ('{}') TO HASH KEYS IN TO CLUSTER SHARDS BETTER. THIS MEANS 0.5 __IS NOT__ COMPATIBLE WITH 0.4 A MIGRATION SCRIPT WILL BE PROVIDED UPON COMPLETION.
- Move all Redis key formating to .format() rather than % syntax
- Make all ids (uid, pid, cid & aid) UUIDs (uuid1) and use the hex repr
- Removing all casts to int for the IDs as format means we do not need to do this, and IDs are no longer integers.
- Move tests from the main Pjuu package to there own package called tests in the repository. This means the Pjuu app can be deployed better.
- General code clean ups from bad mistakes I made when learning Redis.

The goal is too get too 100% coverage and remove any possible race conditions. We are also planning on a move to Redis cluster in the future so a lot of the Redis design is changing.