Added pages feature.
Non-dynamic pages can now be defined in pjuu/pages.py. These should work
when a user is or is not logged in. These are there for our about, terms
of service and privacy policy pages.

Unit tests have been added but there IS NO CONTENT on these pages as of
yet.