Added icon to alerts which are new.
This checks to see if the alert timestamp is newer than the last time
checked and appends an icon. As the alerts are pulled in if they are
newer a new property is applied to the alert item `new`.

Also changed the wording when hidding alerts as you are not actually
deleting them.

Fixed the unit tests and added in some code to keep 100% coverage.