Work started on MongoDB integration...
There had been an awful lot of stuff changed. The code should be simpler with the introduction on PyMongo.

Non of the packages has been fully converted yet. Auth is probably the closest to being done. All the packages are tightly intergrated so all will need updating.

The plan at the moment is to keep lists such as the feed, followers, followees, votes, subscribers and all alerts inside Redis.

Posts, comments and users will be stored inside Mongo. This also allows us to remove the lists of posts and comments as we can lookup the uid inside each hash. This will mean there should be less issues with data deletion and insertion.

Please note, this branch WILL NOT PASS TESTS.