Reworked a lot more stuff...
This still will not pass unit tests, it is actually far from it!

Reworked a lot of functions within posts and users to try and make our way towards fixing issue #19 and issue #20.

Posts and comments are now the same thing so we only have one write path. There is a difference between these. Comments now have a new field called 'reply_to' which should be an ID of the post. Some of the fields have also been renamed within the hash. uid is now user_id for example. This renaming will make its way in to the users collection also in the future.

A lot of parameters have been renamed in a similar fashion to the above. 'uid' are becoming 'user_id' and 'cid's are being replaced with 'reply_id'.

Please note the work on the MongoDB rewrite / full code refactor is still in its infancy. I would like to review all processes.