Code re-org! Moved views in to Blueprints...
The issues I have been facing can be worked around with a bit of code re-org and the use of Blueprints. This seems to have stopped a lot of the issues with working outside the application context.

This is the initial work on this and there is much to do. Hopefully we will be able to removed the tests from there app_context and start with a fresh slate everytime amonst other things.

PLEASE NOTE: URLS ARE NOW PRECEDED WITH THERE BLUEPRINT, EG. 'signin' BECOMES 'auth.signin' THIS SHOULD HELP WITH READING INSIDE TEMPLATES.