Fixed stupid mistake in indexes for Mongo.
What I meant was 'sparse' and 'spacial'. Secondly after refering to the docs again I don't need it because I do want to index the 'null' values.

Please note this file may move in to scripts as it only need to be run once per deploy, this needs some more thinking.