Added an ensure_indexes script...
This is called from the same location as run_server and run_tests so that we can run it everytime we deploy.

Also fixed the test suite to delete the mongo database after each test and not remove indexes and collections. This keeps your machine free of the `pjuu_testing` database.