Fixed `comment_count` on posts when deleting accounts.
The issue with deleting accounts is that I can not call ``pjuu.posts.backend.delete_post`` because it causes a cyclical import. I can't more the commands for getting ``user_id``s from auth as that is where is belongs.

I've copied the code from ``delete_post`` and ``delete_post_replies`` in to the ``delete_account`` function. I WILL FIX THIS LATER. There will be a cleaner place to put these things to stop this happening.

Also added tests to ensure this is happening.