Added Gravatar back to views and to the posts dicts from MongoDB.
I removed this because I was going to add a wrapper view to get this data from your username. I have decided to hold of on this because the issue will be fixed when image uploads in. We may let you choose Gravatar or an uploaded image but the URL will be the same and it will be a background issue.

I replaced these with your username just to be able to get through testing. I only remembered when I ran Pjuu through the Werkzeug profiler and saw hundreds of requests.