Added pagination size change.
This can be set in the profile settings. The actual types of pagination
may need tweaking but this will be easy now.

Added tests and ensure that everything that returns a pagination can
handle these custom settings. Everything is handled in the front end to
do with the user and the backend has been extended to accept pagination
sizes. It will still fall back to the settings `app.config` if these are
not set.