Fix issues with avatar caching and pillow resizing.
- Pillow resize was added rather than thumbnail as it was keeping aspect
  rather than making an odd sized avatar exactly 96x96.
- Removed the need to be logged in to Pjuu to see uploaded content. This
  allows caching at a web tier such as Nginx, Apache, etc.
- So that upload URLs are not the same name as there assoicated objects
  ID. Each upload is assigned a new UUID seperate to that of the post or
  user.

This has caused a load of changes to templates as now, instead of the
users e-mail being pulled in to each post the avatar is now brought in
instead.

This has opened up the ground work to a load more optimisations which
can/need to be made.

Updated avatar migration to correct image sizes.