Hide inactive users from search and tagging.
This was a little tougher than first thought as it needed changing at
the auth app level. Have had to add a flag to allow non-active users to
still show for things such as password reset.

Also show activation status in the dashboard and increased the user list
to 10 users from 5.

As a side. When working on JS stuff earlier I left in a 5 second timeout
on new alerts checks. This has increased load on `pjuu.com` and has now
been returned to 30s.

Closes #191