All images inc animated GIFs have a PNG preview...
Animated GIFs will not be loaded unless you press the `GIF` button.

jQuery GIF player is being used to switch between the PNG preview and
the actual animated GIF.

Tests have been tweaked and every where the `process_upload` is. This
now returned 2 arguments filename and animated_filename.

The templated now also add a `gif` class to an image if it is an
animated gif.