Added Celery and implemented tasks for mail and... (#315)
* Added Celery and implemented tasks for mail and...
pushing new posts to users feeds.

The new celery_app.py is used to initialize the celery workers. It have
been removed from coverage like wsgi.py.

The settings for it DO NOT WORK at this point in time. We need to set
the broker before we create the flask app so we can't use the same
config file as we don't know where they may come from. Going to work on
this now, may be a good to revisit this.

* Missed celery requirement.

* Added new configurator. Reads Flask config before app initialized.

We can then use the same config for initializing the Celery broker and
the Flask application.