Gunicorn/Gevent is now the development server...
CherryPy changes a lot and we _only_ use it as a development server so
it is not worth the extra dependency. We also have no way of testing
Gunicorn so this will give it more time in development.

The run_server.py script was removed in favour of dev_server.py as we
only need to create a debug WSGI app for Gunicorn.