Arm builds (#396) & Single Dockerfile (#397)
Building on Alpine is slow when using buildx on Github actions. There is a discussion to support Wheel builds for Musl/Alpine: https://discuss.python.org/t/wheels-for-musl-alpine/7084/34

We have moved to python:3.8-slim-buster as a base image for the moment.

This commit also adds support for building arm64 images. It also doesn't run the push on PR and only runs it on master branch pushes.

Closes #397
Closes #396